backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# API Documentation
drf-spectacular==0.26.5

# Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Handling
Pillow==10.0.1
python-magic==0.4.27

# AI Integration
openai==1.3.5

# Notifications
sendgrid==6.10.0
twilio==8.10.0

# Video Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.109.0

# Utilities
python-decouple==3.8
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
factory-boy==3.3.0
faker==19.12.0

# Validation
cerberus==1.3.5
marshmallow==3.20.1

# Monitoring
sentry-sdk[django]==1.38.0